requirements.txt
flask
transformers
torch
python-docx
PyPDF2
